requirements.txt
openai>=1.0.0
anthropic>=0.39.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0
deep-translator>=1.11.0
pyyaml>=6.0
gpt-batch>=0.1.0
